# Reduce transpilation on Node.js >=4