--- a/lib/babel-config.js
+++ b/lib/babel-config.js
@@ -4,6 +4,7 @@ var chalk = require('chalk');
 var figures = require('figures');
 var convertSourceMap = require('convert-source-map');
 var objectAssign = require('object-assign');
+var semver = require('semver');
 var colors = require('./colors');
 
 function validate(conf) {
@@ -40,9 +41,13 @@ function lazy(initFn) {
 }
 
 var defaultPresets = lazy(function () {
+	var esPreset = semver.satisfies(process.version, '>=4') ?
+		'babel-preset-es2015-node4' :
+		'babel-preset-es2015';
+
 	return [
 		require('babel-preset-stage-2'),
-		require('babel-preset-es2015')
+		require(esPreset)
 	];
 });