--- a/lib/caching-precompiler.js
+++ b/lib/caching-precompiler.js
@@ -76,7 +76,8 @@ CachingPrecompiler.prototype._transform = function (code, filePath, hash) {
 CachingPrecompiler.prototype._createTransform = function () {
 	var salt = packageHash.sync(
 		[require.resolve('../package.json')].concat(babelConfigHelper.pluginPackages),
-		JSON.stringify(this.babelConfig)
+		JSON.stringify(this.babelConfig),
+		process.version.split('.')[0]
 	);
 
 	return cachingTransform({