@sindresorhus 2016-10-07
@novemberborn Does the caching take into account `process.version`?