@novemberborn 2016-10-07
@sindresorhus 

> Does the caching take into account `process.version`?

No. Would have to [add that in `caching-precompiler.js`](https://github.com/avajs/ava/blob/16b7e538d2061dbab363087be4babd42e4549aec/lib/caching-precompiler.js#L79).