@sindresorhus 2016-10-07
@novemberborn Done. I went with only salting per major Node.js version. Don't want the cache to expire just because the user did a security upgrade.