@sindresorhus 2016-10-09
Thanks for reviewing :)